+++ randovania/layout/preset_migration.py
@@ -1161,6 +1161,15 @@ def _migrate_v84(preset: dict) -> dict:
     return preset
 
 
+def _migrate_v85(preset: dict) -> dict:
+    if preset["game"] == "am2r":
+        config = preset["configuration"]
+        config["first_suit_dr"] = 50
+        config["second_suit_dr"] = 75
+
+    return preset
+
+
 _MIGRATIONS = [
     _migrate_v1,  # v1.1.1-247-gaf9e4a69
     _migrate_v2,  # v1.2.2-71-g0fbabe91
@@ -1246,6 +1255,7 @@ _MIGRATIONS = [
     _migrate_v82,
     _migrate_v83,
     _migrate_v84,
+    _migrate_v85,  # am2r configurable DR
 ]
 CURRENT_VERSION = migration_lib.get_version(_MIGRATIONS)